requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
contextlib2==21.6.0
Django==5.0.7
django-extensions==3.2.3
django-ninja==1.2.2
django-ninja-extra==0.21.2
injector==0.22.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
